Added early drafts of Gatherer's Jar, Keeper's Crock, Settler's Pot and Clay Cauldron
Added model fiels,
Added base class,
Added datagen to support waterlogged and rotation,
I still need to design up the display positions for each of them.
Currently these blocks do not have any functionality and are not yet obtainable.
OMGOSH minecraft's json format is sooo anoying for models. its one thing to model in it, its another to fix issues with rotatino in block states because it falls back to the defaul quietly. Just a log message, is that too much to ask haha