Added particle effects to knapping in hand
This took a little bit more effort because the player's hands are different depending on if you are first person and yourself or everyone else. 1st person needs your head rot and everything else (f5 and other players) need bodyrot. To do this, I needed to be able to send out server messages to everyone but the actioning player like sounds can.